pyproject.toml
[project]
name = "deep_research_from_scratch"
version = "0.1.0"
description = "Build a deep research agent from scratch"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=1.0.0",
    "langgraph-cli[inmem]",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain[google-genai]",
    "langchain-anthropic>=1.0.0",
    "langchain_community>=0.4",
    "langchain_tavily>=0.2.12",
    "langchain_mcp_adapters>=0.1.11",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
    "tavily-python>=0.5.0",
    "pandas>=2.3.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.36",
    "pydantic-settings>=2.6.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.2.0",
    "pgvector>=0.3.6",
    "psycopg[binary]>=3.2.3",
    "alembic>=1.14.0",
    "pypdf>=5.1.0",
    "reportlab>=4.2.5",
    "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
